Import normalizeHeightmap function and apply it to heightmap
The main.tsx file was modified to import a new function, normalizeHeightmap, from the "tytonic" module. This function is then applied to the heightmap variable, with values from 0.0 to 1.0, allowing for better control of the range of height values. Additionally, a console log was added to provide feedback on the process.