Add noise generation functionality to Tytonic
This commit introduces a new file, `noise.ts`, to the Tytonic assembly directory. This file contains the functionality for creating noise data to be used in the rest of the project. Changes include definitions for noise constants, generation of random float and integer numbers, and a new `Noise` class, all of which help in creating noise-based data. In addition, the export statement in `index.ts` file has been updated to include this new file.