Implement noise generation and refactor variables in main.tsx
The main.tsx file received a set of modifications including the importation of newNoise function from the "tytonic" library and definitions of the NOISE_DEFAULT_HURST and NOISE_DEFAULT_LACUNARITY constants. In addition to adding noise generation, the commit also involves refactoring variables such as "heightmap" renamed to "hm" and creation of "hmOf" and "valuesOfHm" variables. These changes are intended to facilitate handling of these variables in the future.